% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_plot.R
\name{base_plot}
\alias{base_plot}
\title{Generates a basic plot}
\usage{
base_plot(x, y)
}
\arguments{
\item{x}{Independent Variable}

\item{y}{Dependent Variable}
}
\value{
ggplot2 object
}
\description{
Provides the framework for a visualization
}
\examples{
x <- rnorm(n = 100, mean = 0, sd = 1)
y <- rnorm(n = 100, mean = 3, sd = 1)

}
