% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_nmb.R
\name{ci_nmb}
\alias{ci_nmb}
\title{Computes a matrix of net incremental benefits}
\usage{
ci_nmb(df, lambdas, B = 1000, conf_level = 0.95)
}
\arguments{
\item{df}{a data frame containing at least columns A, Z, and Y}

\item{lambdas}{list of lambda values, where lambda is an upper bound on available resources, expressed in terms of currency}

\item{B}{number of bootstrap repetitions, defaults to 1000}

\item{conf_level}{the desired nominal coverage of the confidence interval, defaults to 95%}
}
\value{
list of: matrix of dimension B x length(lambdas) of bootstrapped net monetary benefits, vector of size length(lambdas) of lower CI bounds, vector of size length(lambdas) of upper CI bounds
}
\description{
Computes (via bootstrapping for each value of lambda) a matrix (of dimension B X L, where B is the number of desired bootstrap replications and L is the number of lambda constraints under consideration) of incremental net monetary benefits
}
\examples{
B <- 1000
conf_level <- 0.95
A <- rbinom(N, 1, 0.5)
Z <- rnorm(N, 50 + 4.5 * A, 4)
Y <- rnorm(N, 100 + 2 * A, 5)
lambdas <- c(0:1000)/500
df <- prep_data(A,Z,Y)
ci_nmb(df, lambdas, B, conf_level)

}
