% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ci_bounds.R
\name{create_ci_bounds}
\alias{create_ci_bounds}
\title{Returns a list of two vectors}
\usage{
create_ci_bounds(bens, conf_level = 0.95)
}
\arguments{
\item{bens}{a matrix of incremental benefits having dimension B x L, where B is the number of bootstrap replications and L is the number of lambdas under consideration}

\item{conf_level}{a value between 0 and 1 specifying the percentage of confidence desired}
}
\value{
a list of two vectors, upper and lower bounds for each lambda
}
\description{
This function returns upper and lower confidence interval bounds for each value of lambda, contained in a list of two vectors.
}
\examples{
B <- 1000
N <- 20000
A <- rbinom(N, 1, 0.5)
Z <- rnorm(N, 50 + 4.5 * A, 4)
Y <- rnorm(N, 100 + 2 * A, 5)
lambdas <- c(0:1000)/500
df <- data.frame(A,Z,Y)
bens <- boot_ben(df, lambdas, N)
create_ci_bounds(bens, conf_level = 0.90)
}
