% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2gp.R
\name{plot_2gp}
\alias{plot_2gp}
\title{Generates a basic second-generation acceptability curve}
\usage{
plot_2gp(
  lambdas,
  null_scale = 0.15,
  confidence_interval_lows,
  confidence_interval_highs,
  p_vals,
  tol = 1e-06
)
}
\arguments{
\item{lambdas}{a vector of budget constraints}

\item{null_scale}{a scalar controlling the length of the null interval}

\item{confidence_interval_lows}{a vector of the same length as \code{lambdas}, holding the lower CI bound of the net monetary benefits, often coming from the output of \code{ci_nmb()}}

\item{confidence_interval_highs}{a vector of the same length as \code{lambdas}, holding the upper CI bound of the net monetary benefits, often coming from the output of \code{ci_nmb()}}

\item{p_vals}{a vector of the same length as \code{lambdas} holding the second-generation p-value for each lambda, often coming from the output of \code{compute_2gp()}}

\item{tol}{very small number, how far from the second-generation p-value benchmarks you're willing to accept}
}
\value{
a second generation acceptability curve, a ggplot2 object
}
\description{
Uses the confidence interval method and second-generation p-values to evaluate and display conclusions about treatment optimality
}
