% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepares user's data for use in later functions}
\usage{
prep_data(A, Y, Z)
}
\arguments{
\item{A}{a binary vector listing treatment types for each unit of observation (where 0 is the null and 1 is the alternative)}

\item{Y}{a numeric vector listing the cost of the given intervention for each unit of observation}

\item{Z}{a numeric vector listing the effect size of the given intervention for each unit of observation}
}
\value{
data frame consisting of the columns A,Y,Z with appropriate column names
}
\description{
Given a set of vectors, we construct a properly named data frame for use in other package functions
}
