% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visegen.R
\docType{package}
\name{visegen}
\alias{visegen-package}
\alias{visegen}
\title{Creates Visualizations with Second Generation P-Values}
\description{
This section is under construction.
}
\author{
Ashley E. Mullan \email{ashleymullan2000@gmail.com}
}
