% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ben.R
\name{boot_ben}
\alias{boot_ben}
\title{Computes a matrix of net incremental benefits}
\usage{
boot_ben(df, lambdas, B = 1000, N)
}
\arguments{
\item{df}{a data frame containing at least columns A, Z, and Y}

\item{lambdas}{list of lambda values, where lambda is an upper bound on available resources, expressed in terms of currency}

\item{B}{number of bootstrap repetitions, defaults to 1000}

\item{N}{sample size}
}
\value{
matrix of dimension B x length(lambdas)
}
\description{
Computes (via bootstrapping for each value of lambda) a matrix (of dimension B X L, where B is the number of desired bootstrap replications and L is the number of lambda constraints under consideration) of incremental net monetary benefits
}
\examples{
B <- 1000
N <- 20000
A <- rbinom(N, 1, 0.5)
Z <- rnorm(N, 50 + 4.5 * A, 4)
Y <- rnorm(N, 100 + 2 * A, 5)
lambdas <- c(0:1000)/500
df <- data.frame(A,Z,Y)
boot_ben(df, lambdas, N)

}
