% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_nmb.R
\name{ci_nmb}
\alias{ci_nmb}
\title{Computes confidence intervals for incremental net monetary benefits}
\usage{
ci_nmb(df, lambdas, B = 1000, conf_level = 0.95)
}
\arguments{
\item{df}{a data frame containing at least columns A, Z, and Y}

\item{lambdas}{vector of budget constraints}

\item{B}{number of bootstrap repetitions, defaults to 1000}

\item{conf_level}{the desired confidence level, defaults to 95\%}
}
\value{
\item{lower_bounds}{a vector of size length(\code{lambdas}) of the lower bounds of the confidence intervals for each value of lambda}
\item{upper_bounds}{a vector of size length(\code{lambdas}) of the upper bounds of the confidence intervals for each value of lambda}
\item{bNMB}{a matrix of dimension B x length(\code{lambdas}) of the bootstrapped net monetary benefits}
}
\description{
Computes (via bootstrapping for each value of lambda) a matrix (of dimension B X L, where B is the number of desired bootstrap replications and L is the number of lambda constraints under consideration) of incremental net monetary benefits
}
\examples{
#boostrapping parameters
B <- 1000
conf_level <- 0.95
#generate data
A <- rbinom(N, 1, 0.5)
Z <- rnorm(N, 50 + 4.5 * A, 4)
Y <- rnorm(N, 100 + 2 * A, 5)
#prepare data
df <- prep_data(A,Z,Y)
#set constraints
lambdas <- c(0:1000)/500
ci_nmb(df, lambdas, B, conf_level)
}
