% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2gp.R
\name{plot_2gp}
\alias{plot_2gp}
\title{Generates a basic plot}
\usage{
plot_2gp(
  lambdas,
  null_scale = 0.15,
  confidence_interval_lows,
  confidence_interval_highs,
  p_vals,
  tol = 0.01
)
}
\arguments{
\item{x}{Independent Variable}

\item{y}{Dependent Variable}
}
\value{
ggplot2 object
}
\description{
Provides the framework for a visualization
}
\examples{
x <- rnorm(n = 100, mean = 0, sd = 1)
y <- rnorm(n = 100, mean = 3, sd = 1)

}
